added helper and ignored all files in `pa11y` folder for eslint
